Fixes issue w/ missing menu icon.
NSUserDefaults was wiped when user was logged out, so setting the NSUserDefaults in the App Delegate resulted in a bug. This fixes it so when it restores the NSUserDefaults in the event the user is logged out, it's fine. This is why it showed up sometimes and not others (after a restart of the app).